made regex to grab placeholder files

+++ src/methods/replace.js
@@ -1,9 +1,10 @@
 import fs from 'fs'
+import glob from 'glob'
 import path from 'path'
 
 import { formatString } from '../utils'
 
-const notLocalhostPlaceholderFiles = [
+const dockerPlaceholderfiles = [
   'build.sh',
   'controller.yaml',
   'deploy.sh',
@@ -11,6 +12,18 @@ const notLocalhostPlaceholderFiles = [
   'push.sh',
   'run.sh',
   'service.yaml'
+]
+
+const notLocalhostPlaceholderFiles = [
+  /.*push.sh$/,
+  /.*controller.sh$/,
+  /.*build.sh$/,
+  /.*controller.yaml$/,
+  /.*deploy.sh$/,
+  /.*Dockerfile$/,
+  /.*push.sh$/,
+  /.*run.sh$/,
+  /.*service.yaml$/
 ]
 const templatePrefix = '_p_'
 
@@ -20,115 +33,36 @@ export function replace () {
 }
 
 export function replaceProject () {
-  const { project } = this
+  const { program, project } = this
   // boilerplate
-  this.setAllTypesAndServersEnvironment()
-  this.replacePlaceholderFiles()
+  program.method = 'replacePlaceholderFiles'
+  program.methods = null
+  this.mapInTypesAndServers()
   // info
   this.consoleInfo(`Your ${project.package.name} project was successfully replaced!`)
 }
 
 export function replacePlaceholderFiles () {
-  const { program } = this
-  this.setAllTypesAndServersEnvironment()
-  program.image = undefined
-  program.method = null
-  program.methods = [
-    'service.yaml',
-    'controller.yaml',
-    'client_secret.json',
-    'guwsgi.ini',
-    'uwsgi.ini',
-    'guwsgi.ini'
-  ].map(file => {
-    return {
-      folder: 'config',
-      file: file
-    }
-  }).concat([
-    'build.sh',
-    'deploy.sh',
-    'install.sh',
-    'push.sh',
-    'run.sh',
-    'start.sh'
-  ].map(file => {
-    return {
-      folder: 'scripts',
-      file: file
-    }
-  })).concat([
-    'Dockerfile'
-  ].map(file => {
-    return {
-      folder: 'server',
-      file: file
-    }
-  })).map(newProgram => () => {
-    Object.assign(program, newProgram)
-    this.replacePlaceholderFile()
-  })
-  this.mapInTypesAndServers()
-}
-
-export function replacePlaceholderFile () {
-  this.checkProject()
-  const { backend, program, project, run, server, type } = this
+  // unpack
+  const { backend, project, program, type, run, server } = this
   // connect if no port was set here
   if (type.name !== 'localhost' && typeof run.port === 'undefined') {
     this.connect()
   }
-  // check
-  if (!backend || !run || !server || !type ||
-    (type.name === 'localhost' && notLocalhostPlaceholderFiles.includes(program.file))
-  ) { return }
-  // set the file name
-  let installedFileName = program.file
-  if (type) {
-    installedFileName = `${type.name}_${installedFileName}`
-  }
-  // look first if there is no specific <type>_<script> in a certain template
-  let templateFileDir = null
-  // get all templates
-  const allTemplateNames = this.getAllTemplateNames()
-  const foundTemplateName = allTemplateNames.find(templateName => {
-    const templateDir = path.join(project.nodeModulesDir, templateName)
-    const templateServerDir = path.join(templateDir, 'backend/servers', server.name)
-    const templateFolderDir = program.folder === 'server'
-    ? templateServerDir
-    : path.join(templateServerDir, program.folder)
-    let templateFileName = installedFileName
-    templateFileName = `${templatePrefix}${templateFileName}`
-    templateFileDir = path.join(templateFolderDir, templateFileName)
-    if (fs.existsSync(templateFileDir)) return true
-    // remove the type prefix then to find a general <script> template
-    templateFileName = `${templatePrefix}${program.file}`
-    templateFileDir = path.join(templateFolderDir, templateFileName)
-    return fs.existsSync(templateFileDir)
-  })
-  if (!templateFileDir || typeof foundTemplateName === 'undefined') return
-  const templateFile = fs.readFileSync(templateFileDir, 'utf-8')
-  const installedFolderDir = program.folder === 'server'
-  ? server.dir
-  : path.join(server.dir, program.folder)
-  const installedFileDir = path.join(installedFolderDir, installedFileName)
-  // ok for now if the file already exists and that we are not in the force mode, leave
-  if (fs.existsSync(installedFileDir) && program.force !== 'true') {
-    return
-  }
   // prepare the dockerExtraConfig
   const extraConfig = Object.assign(
     {
-      'DOCKER_HOST': run.host,
-      'PORT': run.port,
       'SITE_NAME': backend.siteName,
       'TYPE': type.name,
-      'URL': run.url,
       'WEB': 'on'
     },
     backend.dockerEnv,
     server.dockerEnv
   )
+  // specify the port if it is already set
+  if (run.port) {
+    extraConfig.PORT = run.port
+  }
   this.dockerExtraConfig = Object.keys(extraConfig)
     .map(key => `ENV ${key} ${extraConfig[key]}`).join('\n')
   this.manageExtraConfig = Object.keys(extraConfig)
@@ -136,8 +70,48 @@ export function replacePlaceholderFile () {
   if (this.manageExtraConfig.length > 0) {
     this.manageExtraConfig = `${this.manageExtraConfig} &&`
   }
-  // info
-  this.consoleInfo(`Let\'s install this placeholder file ${installedFileDir}`)
-  // replace
-  fs.writeFileSync(installedFileDir, formatString(templateFile, this))
+  // for each template replace
+  this.getAllTemplateNames().forEach(templateName => {
+    const templateDir = path.join(project.nodeModulesDir, templateName)
+    const templateServerDir = path.join(templateDir, 'backend/servers', server.name)
+    const templateFileDirs = glob.sync(path.join(templateServerDir, `**/${templatePrefix}*`))
+    templateFileDirs.forEach(templateFileDir => {
+      const dirChunks = templateFileDir.split('/')
+      let installedFileName = dirChunks.slice(-1)[0]
+                                       .replace(templatePrefix, '')
+      // we know that there are some script and config files dedicated to the deploy step
+      // so we don't have actually to write them for the localhost type case
+      if (type.name === 'localhost') {
+        if (notLocalhostPlaceholderFiles.find(notLocalhostPlaceholderFile => {
+          return notLocalhostPlaceholderFile.test(installedFileName)
+        })) {
+          return
+        }
+      } else if (installedFileName.split('_')[0] === 'localhost') {
+        // also we have some specific placeholder files dedicated to localhost
+        // so if we are not to the localhost case we have to leave
+        return
+      }
+      // if we are not in the case of a docker deploy file
+      // we can escape
+      if (typeof backend.helpersByName.docker === 'undefined' &&
+        dockerPlaceholderfiles.includes(installedFileName)) {
+        return
+      }
+      const installedParams = installedFileName.split('_')
+      if (installedParams.length === 2 && installedParams[0] !== type.name ||
+      installedParams.length === 1) {
+        installedFileName = `${type.name}_${installedFileName}`
+      }
+      const installedFileDir = path.join('backend', dirChunks.slice(0, -1)
+                                  .join('/')
+                                  .split('backend')
+                                  .slice(-1)[0], installedFileName)
+      if (fs.existsSync(installedFileDir) && program.force !== 'true') {
+        return
+      }
+      const templateFile = fs.readFileSync(templateFileDir, 'utf-8')
+      fs.writeFileSync(installedFileDir, formatString(templateFile, this))
+    })
+  })
 }
